+++ forge/src/main/java/com/iamkaf/amber/mixin/LivingEntityAfterDamageMixin.java
@@ -127,29 +127,24 @@ public abstract class LivingEntityAfterDamageMixin {
         /*this.amber$armAfterDamage(damage, damage, false);*/
     }
 
-    @Inject(
-            //? if >=1.21.2
-            method = "hurtServer",
-            //? if <1.21.2
-            /*method = "hurt",*/
+    //? if <1.18.1 {
+    /*@Inject(
+            method = "hurt",
             at = @At(
                     value = "INVOKE",
                     target =
                     //? if >=1.17
                     "Lnet/minecraft/world/entity/LivingEntity;hurtCurrentlyUsedShield(F)V"
                     //? if <1.17
-                    /*"Lnet/minecraft/entity/LivingEntity;hurtCurrentlyUsedShield(F)V"*/
+                    /^"Lnet/minecraft/entity/LivingEntity;hurtCurrentlyUsedShield(F)V"^/
             )
     )
     private void amber$fireLegacyShieldBlock(
-            //? if >=1.21.2
-            ServerLevel level,
             DamageSource source,
             float damage,
             CallbackInfoReturnable<Boolean> cir
     ) {
-        //? if <1.18.1 {
-        /*if (!((Object) this instanceof Player player)) {
+        if (!((Object) this instanceof Player player)) {
             return;
         }
 
@@ -159,8 +154,8 @@ public abstract class LivingEntityAfterDamageMixin {
         }
 
         PlayerEvents.SHIELD_BLOCK.invoker().onShieldBlock(player, shield, damage, source);
-        *///?}
-    }
+    }*/
+    //?}
 
     @Inject(
             //? if >=1.21.2